bookstore/requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
certifi==2021.10.8
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.0
cssselect2==0.7.0
distlib==0.3.4
Django==4.2.14
django-crispy-forms==2.3
django-rest-framework==0.1.0
django-xhtml2pdf==0.0.4
djangorestframework==3.15.2
filelock==3.6.0
html5lib==1.1
idna==3.7
lxml==5.2.2
oscrypto==1.3.0
pillow==10.4.0
pipenv==2022.1.8
platformdirs==2.5.1
pycparser==2.22
pyHanko==0.25.1
pyhanko-certvalidator==0.26.3
pypdf==4.3.1
pypng==0.20220715.0
python-bidi==0.6.0
PyYAML==6.0.1
gunicorn==19.7.1
qrcode==7.4.2
reportlab==4.2.2
requests==2.32.3
six==1.16.0
sqlparse==0.4.2
svglib==1.5.1
tinycss2==1.3.0
typing_extensions==4.12.2
tzlocal==5.2
uritools==4.0.3
urllib3==2.2.2
virtualenv==20.13.3
virtualenv-clone==0.5.7
webencodings==0.5.1
xhtml2pdf==0.2.16
